[py-tx] Don't reuse adapters (#361)
Summary
---------

The connect pool is controlled by the adapter, which is closed when the session is closed. If you use new sessions, but reuse the adapter, you'll keep closing the connection pool, which then leads to exceptions.

Fixes: https://github.com/facebook/ThreatExchange/issues/359

Test Plan
---------

Do a fetch on a collaboration with at least 10k items, asplodes before, doesn't asplode afterwards.