Working flask app with dockerfile (#380)
Summary
---
The flask app doesn't do anything but print "Hello, from the threatexchange API".

Adds a Dockerfile, adds flask to requirements and orchestrates everything with uwsgi, one of the canonical webservers writting in python.

Test Plan
---
```
$> cd hma-lite/
$> docker build -t threatexchange-api:latest -f ThreatExchangeAPI.Dockerfile .
$> docker run -d -p 8080:80 threatexchange-api
```

And then localhost:8080 on your browser.

If you want to develop locally, use flask. First,
```
$> pip install -r requirements.txt # ensures you have flask installed. Only run once.
$> FLASK_APP=hmalite.app:app FLASK_ENV=development flask run
```

And go to the URL you are presented with in the console logs.