[hmalite] flask config (#386)
Summary
---------

Adds a simple flask config. I'm not a huge fan of it being untyped, but not sure if it's easy to fix either.

Test Plan
---------

Spent several hours learning various ways to fail to run. The winning strategy ended up being:
1. Code in devserver
2. Test by running flask directly and then curling it 
2. push to fork repo
3. pull from fork repo onto laptop (in git-bash)
4. in powershell, build a docker image
5. run the docker image with `docker run -d -p 5000:80 lite`
6. visit localhost:5000 in browser

I know there are many possible optimizations to the above, but this is the only route that works so far. I am not a good engineer.