[ise][hma-lite] config (#388)
Summary
---------

This is the output of some number of hours of disorganized hacking, and so is not the cleanest diff.

Somewhere in this pile, there is:
1. Handling configs (supported patterns = written as file, overridden with environ)
2. A bug fix for the PDQ index library
3. Moving things around

Test Plan
---------

Finally figured out how to connect from my browser to the devserver:

ssh -L 5000:localhost:5000

will do the trick. In my case, it took me a few hours to learn that something was reserving 5000, and so needed to switch to 5001.
May contain random bugs.