[hma][config] Add subtype helper (#510)
Summary
---------
Adds HMAConfigWithSubtypes and subtype logic.

Well... here's a thing that does a thing. I went in circles on this for a while on decision points that probably didn't matter.

The key thing I hung on is where to do this:
```
     HMAConfig
          |
HMAConfigWithSubtype
          |
   HMAConfigSubtype
```

Or this

```
HMAConfig         HMAConfigWithSubtype
     |               - Returns instances of HMAConfigSubtype
HMAConfigSubtype
```

There are tradeoffs in either direction, but decided to go with former. 

I also made the choice(/mistake) to use metaclass magic to provide a default behavior on class name.

Test Plan
---------

Updated unittest, I believe its comprehensive.