[HMA][MultiPart] Count matches per privacy group (#595)
* terraform: Move datastore definition to its own module

* Add methods to increment count of matches by privacy group

* Add PrivacyGroups to PDQ Match Record

This is information that was getting missed. Unless we have a lookup for all indicator ids, this might get tricky to retain. If an indicator_id (what's currently stored) is present in multiple privacy groups, which privacy groups were matched needs to be retained.

I'm not entirely sure whether the changes in main.tf were required. But without it, I was getting KeyErrors.

* Stream-processing lambda: listens to DDB INSERT and MODIFY and updates counts for matches per privacy group

* terraform: connect stream processing lambda to ddb updates

* Customary mypy fixes (no need to review)

* Revert "Add PrivacyGroups to PDQ Match Record"

This reverts commit f7232a18d25079d5e60e88019e2be85d0f446d65.

Based on review feedback, we're taking a different approach. Reverting,
instead of simply removing the commit allows history to be retained.

* Revert "terraform: connect stream processing lambda to ddb updates"

This reverts commit c9cca15d2eae4e0d115714356ff7a6ec7dccd4ae.

Since we are no longer connecting a stream processing lambda, removing
this commit.

* Second queue on matches SNS topic to increment pg counters

* Prevent test runs on python 3.9; restrict to 3.8

* Address review feedback; stale comments and unnecessary permissions in terraform