Support S3 ObjectCreate events in unified hasher; cleanup pdq-* components (#782)
* nitty: Make usage of prefix consistent across images/

The folder construct is inaccurate and potentially confusing. Using the more obvious and "known" prefix terminology.

* failsafe: prevent accidental deletes of the dynamodb

* Extract primary-bucket out of hashing-data

It is used for so much more, (indexing, fetching etc). It might also be time to reconsider the name for hashing-data module.

This was prompted by a weird dependency cycle in trying to get submission_queue and the bucket refer to each other. This was not strictly necessary, but is a step in the direction.

* Support s3 CreateObject events in Unified Hasher

* Swap out pdq-* functions when creating Cloudwatch dashboard

* Cleanup objects in the pdq-signals module.

It still exists so that terraform can identify the missing artifacts and clean them up.

Once we fix #755, this should not be necessary when removing a module.

* Review Feedback

- doc on why datastore has prevent_destroy lifecycle attribute
- convert SQS policy to terraform data object
- acknowledge that we add notification configs on the data_bucket object within other modules.

Test Plan
----
`terraform apply` did not detect any differences so the policy config was correctly translated.