Support parameterized counts in stream counters (#854)
* Support parameterized counts in stream counters

At this point, only adds support for and test a 'buffer' which receives increments and decrements for both aggregate and parameterized counts. It flushes on-demand (or automatically after every 20 writes).

Next steps would be to identify which parameterized counts we want to maintain and inc them in the lambda's `PipelineTableStreamCounter.update_increments_for_records(...)`

Test Plan
---
Added a unittest with a mock ddb table for counts.

Also verified from a local run of the lambda that counts are incremented.

* fix: use source_table_type instead of source_table_name

Source table name could be confused for the actual table name. Also added copious amounts of documentations explaining what's going on in the lambda itself.

Tested by deploying terraform and doing a submission to observe values change in ddb.

* Avoid autoflushes in CountBuffer

Barrett made a good point about there being no rollbacks in lambda execution. Hence, failures in ddb_stream_counter execution might lead to double counting.

So, instead of writing counts automatically, we only write at the end when a subclass of BaseTableStreamCounter calls buffer.flush()

* Address final review comments

- docstring lint
- more descriptive TODO
- avoid magic constants by moving SKEY_PREFIX into class constant