OMM: Roles and blueprints (#1343)
* [omm] Basic HTML index page, separate status endpoint

The index endpoint ('/') is distinct in function from '/status',
so separate it out. Add a basic HTML page here so that user is
greeted with something human-readable to show they're up and running.

* [omm] Define server roles with Flask blueprints for each role

Define the hasher, matcher and curator roles. Allow them to be enabled
or disabled via config. Separate those role functions out into discrete
Flask "Blueprints" with URL prefixes for easy Layer 7 routing.
Add stubs for a handful of example functions from the README.